Duck grammar V. 0.2.15.7.24.
This update increases the number of LR(1) parser states from 305 to 2796 states.

Feature changes include:

Adding the ++ and -- operators.

Added user classes.

Support for operator-overloading through string named functions.

Implements do-while loops.

Added an optional step parameter to for loops.

Syntax changes include:

Replaced the `function' keyword with `func'. Replaced `end' with `end func' for function definitions.

Replacing the terminating `end' in if-statements with `endif'.

Added an optional `static' keyword for functions and assignments.

Replaced dictionary syntax with `[', `]', as opposed to curly braces (`{', `}') the same as arrays or lists.

Added keywords for `new' and `this'.